Refactored Accordion component to simplify inheritance.
  Prior to this change the Accordion component was wrapped with
react-redux in order to bind it to the application state. Wrapping the
Accordion component this way caused problems when trying to use
component state in subclasses. For instance, it was impossible to fade a
child component as the Accordion scrolled, because you can't pass
component state into the wrapped inner Accordion component. This was
confusing, so it was changed to pull in the application state directly
into the component, which makes subclassing the Accordion work as
expected.