Refactor Link to improve default styling.
  Prior to this change the Link element wrapped React Native's
TouchableHighlight element, this gives a horrible black flash when
tapped in the default styling. To counter this I've changed the wrapping
element to TouchableOpacity, which is a little less flexible but gives
much better default styling. A future change may be necessary where the
consumer is given the option, as well as more flexibility in styling,
but for now this works.